+++ tests/benchmarks/src/ack.rs
@@ -25,7 +25,7 @@ pub struct AckRunConfig {
     pub count: u64,
     pub concurrency: usize,
     pub seed_offset: u64,
-    pub max_fee: u32,
+    pub max_fee: u16,
     pub request_timeout_ms: u64,
     pub fail_on_rejection: bool,
     pub workload: WorkloadConfig,
